:art: Style fix

+++ src/test/gasTests/BuyVault2.sol
@@ -68,7 +68,7 @@ contract gasBuyVault_2ERC20 is Test {
     DebtToken debt;
 
     event log_named_array(string key, address[] val);
-    event log_named_array(string key, uint[] val);
+    event log_named_array(string key, uint256[] val);
 
     address private creatorAddress = address(1);
     address private tokenCreatorAddress = address(2);
